Исправлено преобразование дробных чисел из строки.

+++ Code/ComfortIsland/Helpers/DigitRoundingConverter.cs
@@ -14,7 +14,7 @@ namespace ComfortIsland.Helpers
 
 		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			return double.Parse(value.ToString());
+			return decimal.Parse(value.ToString());
 		}
 
 		public static string Simplify(decimal value)
